+++ CHANGES.rst
@@ -5,6 +5,8 @@ Unreleased
 ----------
 
 * Reduce debug.log noise by avoiding the getinfo RPC
+* Limit the number of plotted mempool transactions to maintain performance
+  under a backlog (currently hardcoded to 5000)
 
 0.1.0 (2015-02-02)
 ------------------